+++ src/components/overlay/overlayModeEnum.ts
@@ -1,0 +1,15 @@
+'use strict';
+
+/**
+ * Enum for the modules supported by the Overlay directive.
+ * 
+ * @ngdoc enum
+ * @readonly
+ * @enum {string}
+ * @name OverlayMode
+ *
+ */
+export enum OverlayMode {
+  light,
+  dark
+}